feat: Contact page gets cloaked string on mount, and moved to TS
