+++ package.json
@@ -39,10 +39,12 @@
     "@next/bundle-analyzer": "^11.0.1",
     "@testing-library/jest-dom": "^5.1.0",
     "@testing-library/react": "^9.4.0",
+    "@types/cookies": "^0.7.7",
     "@types/node": "^14.14.7",
     "@types/nprogress": "^0.2.0",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
+    "@types/stegcloak": "^1.0.0",
     "@types/styled-components": "^5.1.4",
     "@types/styled-system": "^5.1.10",
     "@typescript-eslint/eslint-plugin": "^4.28.5",
